[Update]いいね重複防止の記述追加

+++ app/controllers/public/favorites_controller.rb
@@ -4,13 +4,21 @@ class Public::FavoritesController < ApplicationController
   before_action :get_vision_id, only: [:create, :destroy]
 
   def create
-    favorite = @vision.favorites.new(member_id: current_member.id)
-    favorite.save
+    # 既にいいねが存在する場合は処理をスキップ
+    unless already_favorited?
+      favorite = @vision.favorites.new(member_id: current_member.id)
+      favorite.save
+    end
+    # 必要に応じてリダイレクト先を変更するなどの処理を追加
   end
 
   def destroy
-    favorite = current_member.favorites.find_by(vision_id: @vision.id)
-    favorite.destroy
+    # いいねが存在する場合のみ削除
+    if already_favorited?
+      favorite = current_member.favorites.find_by(vision_id: @vision.id)
+      favorite.destroy
+    end
+    # 必要に応じてリダイレクト先を変更するなどの処理を追加
   end
 
   private
@@ -23,4 +31,8 @@ class Public::FavoritesController < ApplicationController
     @vision = Vision.find(params[:vision_id])
   end
 
+  def already_favorited?
+    # すでにいいねが存在するか確認
+    current_member.favorites.exists?(vision_id: @vision.id)
+  end
 end
